+++ src/i18n/uk-translation.ts
@@ -72,6 +72,7 @@ const ukTranslation = {
     project_tags: 'Кількість тегів може бути від 1 до 5',
     project_summary: 'Короткий опис і мета не може бути пуста або мати більше 150 символів',
     project_description: 'Опис проєкта не може бути пуста або мати більше 512 символів',
+    project_deadline: 'Дата дедлайну збору не може бути більшою за дату дедлайну повернення',
   },
 };
